+++ cpp/oneapi/dal/util/detail/load_graph.hpp
@@ -24,38 +24,12 @@
 #include "oneapi/dal/data/undirected_adjacency_array_graph.hpp"
 #include "oneapi/dal/exceptions.hpp"
 #include "oneapi/dal/util/csv_data_source.hpp"
+#include "oneapi/dal/util/detail/load_graph_service.hpp"
 #include "oneapi/dal/util/load_graph_descriptor.hpp"
 #include "services/daal_atomic_int.h"
 #include "services/daal_memory.h"
 
-namespace oneapi::dal::preview {
-typedef void (*functype)(int i, const void *a);
-}
-
-extern "C" {
-ONEAPI_DAL_EXPORT void _daal_threader_for_oneapi(int n,
-                                                 int threads_request,
-                                                 const void *a,
-                                                 oneapi::dal::preview::functype func);
-}
-
 namespace oneapi::dal::preview::load_graph::detail {
-
-template <typename F>
-inline void threader_func(int i, const void *a) {
-    const F &lambda = *static_cast<const F *>(a);
-    lambda(i);
-}
-
-template <typename F>
-inline ONEAPI_DAL_EXPORT void threader_for(size_t n, size_t threads_request, const F &lambda) {
-    const void *a = static_cast<const void *>(&lambda);
-
-    _daal_threader_for_oneapi((int)n, (int)threads_request, a, threader_func<F>);
-}
-
-ONEAPI_DAL_EXPORT int daal_string_to_int(const char *nptr, char **endptr);
-
 edge_list<std::int32_t> load_edge_list(const std::string &name) {
     using int_t = std::int32_t;
 
@@ -79,7 +53,6 @@ edge_list<std::int32_t> load_edge_list(const std::string &name) {
 template <typename Graph>
 void convert_to_csr_impl(const edge_list<vertex_type<Graph>> &edges, Graph &g) {
     auto layout           = oneapi::dal::preview::detail::get_impl(g);
-    using int_t           = typename Graph::vertex_size_type;
     using vertex_t        = typename Graph::vertex_type;
     using vector_vertex_t = typename Graph::vertex_set;
     using vector_edge_t   = typename Graph::edge_set;
